requirements.txt
mcp
google-play-scraper
requests
openai
google-generativeai
umap-learn
hdbscan
spacy
thefuzz
pydantic
sqlalchemy
pyyaml
jinja2
python-dotenv
tiktoken
emoji
langdetect
groq
google-generativeai

fastapi
uvicorn
python-dotenv
